+++ switch_my_loan/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.7'
+__version__ = '1.0.10-uat'
 
 import frappe
 import os