+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [v1.0.10-uat] - 2022-06-28
+## [v1.0.10-dev] - 2022-06-28
+
+### Added
+
+[notifications addition](https://gitlab.com/atri-tech/atri-maintainers/switch-my-loan/-/merge_requests/47)
+
+## [v1.0.9-dev] - 2022-06-27
+
+### Added
+
+[reports|reason for rejection mandatory|Product wise mobile number should be unique](https://gitlab.com/atri-tech/atri-maintainers/switch-my-loan/-/merge_requests/46)
+
+## [v1.0.8-uat] - 2022-06-06
+## [v1.0.8-dev] - 2022-05-31
+
+### Added
+
+[export right to partner roles|exco role addition|Sales manager can see their granchildren leads|lead allocation sms]
+
 ## [v1.0.7] - 2022-05-27
 ## [v1.0.7-uat] - 2022-05-25
 ## [v1.0.7-dev] - 2022-05-25